Moved variable initialization to utils, and cleaned up some import statements
